+++ pkg/model/gateways/event_based.go
@@ -1,4 +1,3 @@
-// Package gateways provides BPMN gateway implementations.
 package gateways
 
 import (